address feedback for https://github.com/Microsoft/ApplicationInsights-aspnet5/pull/66
