+++ test/FunctionalTestUtils/TelemetryTestsBase.cs
@@ -1,16 +1,50 @@
-﻿
-namespace FunctionalTestUtils.Tests
+﻿namespace FunctionalTestUtils
 {
     using System;
     using System.Linq;
+    using System.Collections.Generic;
+    using System.Net.Http;
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.DataContracts;
     using Xunit;
 
-    public abstract class RequestTelemetryTestsBase : TelemetryTestsBase
+    public abstract class TelemetryTestsBase : IDisposable
     {
-        public RequestTelemetryTestsBase(string assemblyName) : base(assemblyName)
-        { }
+        protected const int TestTimeoutMs = 5000;
+
+        private InProcessServer server;
+        private IList<ITelemetry> buffer = new List<ITelemetry>();
+        private HttpClient client = new HttpClient();
+
+        public TelemetryTestsBase(string assemblyName)
+        {
+            this.server = new InProcessServer(assemblyName);
+            BackTelemetryChannelExtensions.InitializeFunctionalTestTelemetryChannel(buffer);
+        }
+
+        public IList<ITelemetry> Buffer
+        {
+            get
+            {
+                return this.buffer;
+            }
+        }
+
+        public InProcessServer Server
+        {
+            get
+            {
+                return this.server;
+            }
+        }
+
+        public HttpClient HttpClient
+        {
+            get
+            {
+                return this.client;
+            }
+        }
 
         public void ValidateBasicRequest(string requestPath, RequestTelemetry expected)
         {
@@ -18,12 +52,8 @@ namespace FunctionalTestUtils.Tests
             var task = this.HttpClient.GetAsync(this.Server.BaseHost + requestPath);
             task.Wait(TestTimeoutMs);
             var result = task.Result;
-            
-            var items = this.Buffer.Where((item) => { return item is RequestTelemetry; });
-            Assert.Equal(1, items.Count());
-            ITelemetry telemetry = items.First();
-            Assert.IsAssignableFrom(typeof(RequestTelemetry), telemetry);
-            var actual = (RequestTelemetry)telemetry;
+
+            var actual = this.Buffer.OfType<RequestTelemetry>().Single();
 
             Assert.Equal(expected.ResponseCode, actual.ResponseCode);
             Assert.Equal(expected.Name, actual.Name);
@@ -41,17 +71,26 @@ namespace FunctionalTestUtils.Tests
             task.Wait(TestTimeoutMs);
             var result = task.Result;
 
-            var items = this.Buffer.Where((item) => { return item is ExceptionTelemetry; });
-            Assert.Equal(1, items.Count());
-            ITelemetry telemetry = items.First();
-            Assert.IsAssignableFrom(typeof(ExceptionTelemetry), telemetry);
-            var actual = (ExceptionTelemetry)telemetry;
+            var actual = this.Buffer.OfType<ExceptionTelemetry>().Single();
 
             Assert.Equal(expected.Exception.GetType(), actual.Exception.GetType());
             Assert.NotEmpty(actual.Exception.StackTrace);
             Assert.Equal(actual.HandledAt, actual.HandledAt);
             Assert.NotEmpty(actual.Context.Operation.Name);
             Assert.NotEmpty(actual.Context.Operation.Id);
+        }
+
+        public void Dispose()
+        {
+            if (this.server != null)
+            {
+                this.server.Dispose();
+            }
+
+            if (this.client != null)
+            {
+                this.client.Dispose();
+            }
         }
     }
 }
