+++ test/WebApiShimFw46.FunctionalTests/Controllers/ValuesController.cs
@@ -20,10 +20,6 @@ namespace SampleWebAPIIntegration.Controllers
         [HttpGet("{id}")]
         public string Get(int id)
         {
-            if (id == 42)
-            {
-                throw new InvalidOperationException();
-            }
             return "value";
         }
